Move light client to separate mod.
