+++ src/main.rs
@@ -6,28 +6,22 @@ use std::{
 	str::FromStr,
 	sync::{mpsc::sync_channel, Arc},
 	thread, time,
-	time::SystemTime,
 };
 
 use anyhow::{Context, Result};
-use futures_util::{SinkExt, StreamExt};
 use ipfs_embed::{Multiaddr, PeerId};
 use rocksdb::{ColumnFamilyDescriptor, Options, DB};
 use simple_logger::SimpleLogger;
 use structopt::StructOpt;
-use tokio_tungstenite::tungstenite::protocol::Message;
 
-use crate::{
-	data::fetch_cells_from_ipfs,
-	http::calculate_confidence,
-	types::{cell_ipfs_record, cell_to_ipfs_block, Mode},
-};
+use crate::types::Mode;
 
 mod app_client;
 mod client;
 mod consts;
 mod data;
 mod http;
+mod light_client;
 mod proof;
 mod rpc;
 mod sync;
@@ -163,162 +157,21 @@ pub async fn do_main() -> Result<()> {
 	}
 
 	let block_header = rpc::get_chain_header(&rpc_url).await?;
-
 	let latest_block = block_header.number;
-	let rpc_ = rpc_url.clone();
-	let db_2 = db.clone();
-	let ipfs2_ = ipfs.clone();
 
 	// TODO: implement proper sync between bootstrap completion and starting the sync function
 	thread::sleep(time::Duration::from_secs(3));
-	tokio::spawn(async move {
-		sync::sync_block_headers(rpc_.clone(), 0, latest_block, db_2, ipfs2_).await;
-	});
 
-	const BODY: &str = r#"{"id":1, "jsonrpc":"2.0", "method": "chain_subscribeFinalizedHeads"}"#;
-	let urls = rpc::parse_urls(cfg.full_node_ws)?;
-	log::info!("Syncing block headers from 0 to {}", latest_block);
-	while let Some(z) = rpc::check_connection(&urls).await {
-		let (mut write, mut read) = z.split();
-		write
-			.send(Message::Text(BODY.to_string()))
-			.await
-			.context("ws-message(chain_subscribeFinalizedHeads) send failed")?;
+	tokio::task::spawn(sync::sync_block_headers(
+		rpc_url.clone(),
+		0,
+		latest_block,
+		db.clone(),
+		ipfs.clone(),
+	));
 
-		log::info!("Connected to Substrate Node");
-
-		let db_3 = db.clone();
-		let cf_handle_0 = db_3
-			.cf_handle(consts::CONFIDENCE_FACTOR_CF)
-			.context("failed to get cf handle")?;
-		let cf_handle_1 = db_3
-			.cf_handle(consts::BLOCK_HEADER_CF)
-			.context("failed to get cf handle")?;
-
-		while let Some(message) = read.next().await {
-			let data = message?.into_data();
-			match serde_json::from_slice(&data) {
-				Ok(types::Response { params, .. }) => {
-					let header = &params.header;
-
-					// now this is in `u64`
-					let num = header.number;
-
-					let begin = SystemTime::now();
-
-					// TODO: Setting max rows * 2 to match extended matrix dimensions
-					let max_rows = header.extrinsics_root.rows * 2;
-					let max_cols = header.extrinsics_root.cols;
-					if max_cols < 3 {
-						log::error!("chunk size less than 3");
-					}
-					let commitment = header.extrinsics_root.commitment.clone();
-
-					let positions = rpc::generate_random_cells(max_rows, max_cols);
-					log::info!(
-						"Random cells generated: {} from block: {}",
-						positions.len(),
-						num
-					);
-
-					let (ipfs_fetched, unfetched) =
-						fetch_cells_from_ipfs(&ipfs, num, &positions).await?;
-
-					log::info!(
-						"Number of cells fetched from IPFS for block {}: {}",
-						num,
-						ipfs_fetched.len()
-					);
-
-					let rpc_fetched = rpc::get_kate_proof(&rpc_url, num, unfetched).await?;
-
-					log::info!(
-						"Number of cells fetched from RPC for block {}: {}",
-						num,
-						rpc_fetched.len()
-					);
-
-					let mut cells = vec![];
-					cells.extend(ipfs_fetched);
-					cells.extend(rpc_fetched.clone());
-
-					if positions.len() > cells.len() {
-						log::error!("Failed to fetch {} cells", positions.len() - cells.len());
-						continue;
-					}
-
-					let count = proof::verify_proof(
-						num,
-						max_rows,
-						max_cols,
-						cells.clone(),
-						commitment.clone(),
-					);
-					log::info!(
-						"Completed {} verification rounds for block {}\t{:?}",
-						count,
-						num,
-						begin
-							.elapsed()
-							.context("failed to get complete verification")?
-					);
-
-					// write confidence factor into on-disk database
-					db_3.put_cf(&cf_handle_0, num.to_be_bytes(), count.to_be_bytes())
-						.context("failed to write confidence factor")?;
-
-					let conf = calculate_confidence(count);
-					log::info!("Confidence factor for block {}: {}", num, conf);
-
-					// push latest mined block's header into column family specified
-					// for keeping block headers, to be used
-					// later for verifying IPFS stored data
-					//
-					// @note this same data store is also written to in
-					// another competing thread, which syncs all block headers
-					// in range [0, LATEST], where LATEST = latest block number
-					// when this process started
-					db_3.put_cf(
-						&cf_handle_1,
-						num.to_be_bytes(),
-						serde_json::to_string(&header)?.as_bytes(),
-					)
-					.context("failed to write block header")?;
-
-					// Push the randomly selected cells to IPFS
-					for cell in rpc_fetched {
-						if let Err(error) = ipfs.insert(cell_to_ipfs_block(cell.clone())) {
-							log::info!(
-								"Error pushing cell to IPFS: {}. Cell reference: {}",
-								error,
-								cell.reference(num)
-							);
-						}
-						// Add generated CID to DHT
-						if let Err(error) = ipfs
-							.put_record(cell_ipfs_record(&cell, num), ipfs_embed::Quorum::One)
-							.await
-						{
-							log::info!(
-								"Error inserting new record to DHT: {}. Cell reference: {}",
-								error,
-								cell.reference(num)
-							);
-						}
-					}
-					if let Err(error) = ipfs.flush().await {
-						log::info!("Error flushing data to disk: {}", error,);
-					};
-
-					// notify ipfs-based application client
-					// that newly mined block has been received
-					block_tx
-						.send(types::ClientMsg::from(params.header))
-						.context("failed to send block message")?;
-				},
-				Err(error) => log::info!("Misconstructed Header: {:?}", error),
-			}
-		}
+	if let Err(error) = light_client::run(cfg.full_node_ws, db, ipfs, rpc_url, block_tx).await {
+		log::info!("Error running light client: {}", error);
 	}
 
 	// inform ipfs-backed application client running thread
